moved the basic use classes to the main package
i want the basic definitions of each of these classes available to all of the design patterns, saves me the time of rewriting them
