+++ src/main/java/Truck.java
@@ -1,4 +1,4 @@
-package main.java.Builder;
+package main.java;
 
 public class Truck extends Car{
